feat(updated): add useObjectDetails & useObjectDetailsMap & useObjectsList hooks
