+++ src/screens/ObjectDetails/hooks/index.tsx
@@ -1,0 +1,2 @@
+export {useObjectDetails} from './useObjectDetails';
+export {useObjectDetailsStatusBar} from './useObjectDetailsStatusBar';
