+++ src/screens/ObjectDetailsMap/hooks/index.tsx
@@ -1,0 +1,1 @@
+export {useObjectDetailsMap} from './useObjectDetailsMap';
