+++ src/screens/ObjectsList/ObjectsList.tsx
@@ -1,64 +1,21 @@
-import React, {useLayoutEffect, useMemo, useCallback} from 'react';
+import React, {useLayoutEffect} from 'react';
 import {FlatList} from 'react-native';
 import {ObjectCard} from 'molecules';
 import {styles} from './styles';
-import {IProps} from './types';
 
-import {
-  useCategoryObjects,
-  useObjects,
-  useObjectsListAnalytics,
-} from 'core/hooks';
 import {SCREEN_WIDTH} from 'services/PlatformService';
 import {PADDING_HORIZONTAL} from 'core/constants';
-import {IObject} from 'core/types';
-import {debounce, orderBy} from 'lodash';
+import {useObjectsList} from './hooks';
 
 const cardWidth = SCREEN_WIDTH - PADDING_HORIZONTAL * 2;
-export const ObjectsList = ({
-  route,
-  navigation: {setOptions, push},
-}: IProps) => {
+export const ObjectsList = () => {
   const {
-    params: {categoryId, title, objectsIds},
-  } = route;
-
-  const {sendSaveCardEvent, sendSelectCardEvent, sendUnsaveCardEvent} =
-    useObjectsListAnalytics();
-
-  const listData = useCategoryObjects(categoryId);
-
-  const listDataByIds = useObjects(objectsIds || []);
-
-  const navigateToObjectDetails = useCallback(
-    ({id, name, category}: IObject) => {
-      push('ObjectDetails', {objectId: id});
-      sendSelectCardEvent(name, category.name);
-    },
-    [push, sendSelectCardEvent],
-  );
-
-  const sortedListData = useMemo(() => {
-    const data = objectsIds ? listDataByIds : listData;
-    return data ? orderBy(data, [({name}) => name.toLowerCase()], 'asc') : null;
-  }, [listData, listDataByIds, objectsIds]);
-
-  const navigateToObjectDetailsDebounced = useMemo(
-    () =>
-      debounce(navigateToObjectDetails, 300, {leading: true, trailing: false}),
-    [navigateToObjectDetails],
-  );
-
-  const sendIsFavoriteChangedEvent = useCallback(
-    ({name, category}: IObject, nextIsFavoriteStatus: boolean) => {
-      if (nextIsFavoriteStatus) {
-        sendSaveCardEvent(name, category.name);
-      } else {
-        sendUnsaveCardEvent(name, category.name);
-      }
-    },
-    [sendSaveCardEvent, sendUnsaveCardEvent],
-  );
+    title,
+    sortedListData,
+    navigateToObjectDetailsDebounced,
+    sendIsFavoriteChangedEvent,
+    setOptions,
+  } = useObjectsList();
 
   useLayoutEffect(() => {
     setOptions({
